# calendar_backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Calendar Backend API",
    "version": "1.0.0",
    "description": "Event CRUD + date-range query APIs with Postgres persistence and in-process notification scheduling."
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health check",
        "description": "Basic health check endpoint.",
        "operationId": "GetHealth",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List events (optionally within a date range)",
        "description": "Returns events; when start/end are provided, returns events overlapping the range. Query params: start, end (ISO-8601).",
        "operationId": "ListEvents",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Create an event",
        "description": "Creates an event and schedules a notification (MVP: 5 minutes before start).",
        "operationId": "CreateEvent",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/events/{id}": {
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Update an event",
        "description": "Updates an existing event and reschedules notifications.",
        "operationId": "UpdateEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Delete an event",
        "description": "Deletes an event and cancels pending notifications.",
        "operationId": "DeleteEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List events (optionally within a date range) (alias)",
        "description": "Compatibility alias for /api/events.",
        "operationId": "ListEventsAlias",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Create an event (alias)",
        "description": "Compatibility alias for /api/events.",
        "operationId": "CreateEventAlias",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/events/{id}": {
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Update an event (alias)",
        "description": "Compatibility alias for /api/events/{id}.",
        "operationId": "UpdateEventAlias",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Delete an event (alias)",
        "description": "Compatibility alias for /api/events/{id}.",
        "operationId": "DeleteEventAlias",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {}
  }
}